Split some modules.
